requirements.txt
Django==4.1.2
django-extensions==3.2.1
flake8==5.0.4
flake8-builtins==2.0.0
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==8.5.0
pep8-naming==0.13.2
python-dotenv==0.21.0
Faker==15.0.0
webargs==8.2.0
crispy-bootstrap5==0.7
django-filter==22.1